[fix] Move ThemingService from libs/common to libs/angular
